Remove JSON-LD, clarify JSON conversion
Fix #78, #80

The JSON-LD direct conversion algorithm is harder to get right, and is
redundant in practice.

Clarify that the "JSON" conversion is explicitly for
`application/microdata+json` and mark it as obsolete but conforming
since it seems to have been removed from current versions of known
implementations.

(And update status)